tools/mtp-registry/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mtp-registry"
version = "0.6.0"
description = "Registry, signing, and approval workflow tooling for MTP artifacts"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Lubor Fedák", email = "lubor@codemoravia.com" },
]
keywords = ["mtp", "registry", "signatures", "approvals", "workflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "click>=8.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
mtp-registry = "mtp_registry.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mtp_registry"]
